gui: redesign ImgListView and rename some modules (#741)
